requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
filelock==3.19.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
setuptools==70.2.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.14.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
